+++ django_mysql/models/__init__.py
@@ -4,8 +4,8 @@ from django_mysql.models.aggregates import (  # noqa
 )
 from django_mysql.models.expressions import ListF, SetF  # noqa
 from django_mysql.models.query import (  # noqa
-    ApproximateInt, SmartChunkedIterator, SmartIterator, pt_visual_explain,
-    QuerySet, QuerySetMixin
+    add_QuerySetMixin, ApproximateInt, SmartChunkedIterator, SmartIterator,
+    pt_visual_explain, QuerySet, QuerySetMixin
 )
 from django_mysql.models.fields import (  # noqa
     Bit1BooleanField, DynamicField, ListCharField, ListTextField,